refactor: payment_screen 분리 — 썸네일 스트립·스타일 카드·선택 이펙트 추출 (JKLI-214)
동작 변경 없는 순수 이동. 984줄이던 화면을 결제 오케스트레이션 중심으로 줄이고,
자기 상태(ScrollController·페이드 플래그)를 가진 모듈을 위젯으로 독립:

- widgets/back_photo_thumbnail_strip.dart — 한화 썸네일 스트립.
  (cards, selectedIndex, onSelect) 인터페이스, 스크롤 상태를 스스로 관리
- widgets/back_photo_style_cards.dart — 풀이미지/라벨 스타일 카드 2장
- widgets/selection_effects.dart — SelectionMotion/Transition/Ring/CheckBadge
  (2단계 커스텀 확장 시 재사용 대비 공용화)
- widgets/back_photo_images.dart — CoverNetworkImage, BackPhotoPlaceholder
